+++ packages/lumx-vue/src/components/message/index.ts
@@ -1,0 +1,8 @@
+import { CLASSNAME, MessageProps } from '@lumx/core/js/components/Message';
+
+import Message from './Message.vue';
+
+Message.className = CLASSNAME;
+
+export { Message };
+export type { MessageProps };