+++ CHANGELOG.md
@@ -11,6 +11,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   `@lumx/vue`:
     -   Create the `Heading` component
+    -   Create the `Message` component
+
+### Changed
+
+-   `@lumx/core`:
+    -   Moved `Message` from `@lumx/react`
 
 ## [4.2.0][] - 2026-02-02